+++ Runtime/Assembly.cs
@@ -3,5 +3,6 @@ using UnityEngine.Scripting;
 
 [assembly: AlwaysLinkAssembly]
 #if UNITY_EDITOR
+[assembly: InternalsVisibleTo("Unity.XR.ARKit.Editor.Tests")]
 [assembly: InternalsVisibleTo("Unity.XR.ARKit.Editor")]
 #endif