+++ Documentation~/whats-new.md
@@ -9,6 +9,7 @@ The main updates in this release include:
 
 **Added**
 
+- Added support for temporal smoothing of the environment depth image. When enabled and supported, this plugin now provides the environment depth image using [ARFrame.smoothedSceneDepth](https://developer.apple.com/documentation/arkit/arframe/3674209-smoothedscenedepth?language=objc) instead of [ARFrame.sceneDepth](https://developer.apple.com/documentation/arkit/arframe/3566299-scenedepth?language=objc).
 - [ARKit reference objects](xref:UnityEngine.XR.ARKit.ARKitReferenceObjectEntry) now store their data in the entry asset, which allows them to be used in [AssetBundles](xref:UnityEngine.AssetBundle).
 
 **Updated**